refactor: upd towers data
